# openapi/omnora.v1.yaml
openapi: 3.1.0
info:
  title: Omnora REST API
  version: 0.1.0
  summary: OpenAPI contract for the first Omnora version.
  description: >
    Contract aligned to the runnable Omnora REST surface. Mount content access is a
    pure account whitelist: effective permission is the lower of current space
    membership and the current mount grant. System administrator status controls
    the control plane and never substitutes for a mount grant. Stale starter paths
    are marked x-omnora-stale and should not be used by new clients.
...
servers:
  - url: /api/v1
...
security:
  - cookieAuth: []
...
paths:
...
  /admin/mounts/{mountId}:
    get:
      tags: [Admin]
      summary: Read mount settings, grants, and eligible space members.
      description: >
        Returns the canonical name/health fields and compatibility aliases
        displayName/status. effectivePermission is computed by the server.
      x-omnora-route-group: rest
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/MountId"
      responses:
        "200":
          description: Mount access detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMountDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: Update mutable mount settings.
      description: Disabling public sharing revokes active shares for this mount; switching to read-only cancels active uploads.
      x-omnora-route-group: rest
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/MountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminMountRequest"
      responses:
        "200":
          description: Updated mount detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMountDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Admin]
      summary: Soft-delete a mount and invalidate its derived capabilities.
      x-omnora-route-group: rest
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/MountId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                deleteData:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Mount deletion result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAdminMountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AI Token or session-bound API token
    cookieAuth:
      type: apiKey
      in: cookie
      name: omnora_session
  parameters:
...
    MountId:
      name: mountId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/OpaqueId"
...
  responses:
    BadRequest:
      description: The request body or parameters are invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication failed or credentials are not currently valid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The current authorization formula denies this operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The object does not exist or is not visible to the caller.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The requested state conflicts with current system state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    OpaqueId:
      type: string
      minLength: 8
      maxLength: 128
      description: Opaque, non-host-path identifier.
    Error:
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - unauthorized
            - forbidden
            - route_group_disabled
            - not_found
            - invalid_cursor
            - quota_exceeded
            - rate_limited
            - mount_identity_unverifiable
            - mount_conflict
            - mount_offline
            - readonly_mount
            - object_changed
            - preview_not_supported
            - share_password_required
            - share_invalid
            - upload_conflict
        detail:
          type: string
...
    AdminMountDetail:
      type: object
      required: [id, spaceId, spaceName, name, displayName, rootPath, kind, mode, indexEnabled, allowPublicShares, health, status, tone, grantCount, grants, eligibleAccounts]
      properties:
        id:
          $ref: "#/components/schemas/OpaqueId"
        spaceId:
          $ref: "#/components/schemas/OpaqueId"
        spaceName:
          type: string
        name:
          type: string
          description: Canonical display name used by the storage UI.
        displayName:
          type: string
          description: Compatibility alias of name.
        rootPath:
          type: string
          description: Configured absolute container path. This control-plane field does not grant content access.
        kind:
          type: string
          enum: [managed, external]
        mode:
          type: string
          enum: [read_only, read_write]
        indexEnabled:
          type: boolean
        allowPublicShares:
          type: boolean
        health:
          type: string
          description: Canonical mount state used by the storage UI.
        status:
          type: string
          description: Compatibility alias of health.
        tone:
          type: string
        grantCount:
          type: integer
          minimum: 0
        grants:
          type: array
          items:
            $ref: "#/components/schemas/AdminMountGrant"
        eligibleAccounts:
          type: array
          items:
            $ref: "#/components/schemas/AdminMountEligibleAccount"
    AdminMountGrant:
      type: object
      required: [accountId, email, displayName, spacePermission, permission, effectivePermission]
      properties:
        accountId:
          $ref: "#/components/schemas/OpaqueId"
        email:
          type: string
        displayName:
          type: string
        spacePermission:
          $ref: "#/components/schemas/MountPermission"
        permission:
          $ref: "#/components/schemas/MountPermission"
        effectivePermission:
          allOf:
            - $ref: "#/components/schemas/MountPermission"
          description: Server-computed minimum of spacePermission and permission.
    AdminMountEligibleAccount:
      type: object
      required: [accountId, email, displayName, spacePermission, permission]
      properties:
        accountId:
          $ref: "#/components/schemas/OpaqueId"
        email:
          type: string
        displayName:
          type: string
        spacePermission:
          $ref: "#/components/schemas/MountPermission"
        permission:
          type: "null"
...
    MountPermission:
      type: string
      enum: [viewer, editor, manager]
...
    UpdateAdminMountRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        displayName:
          type: string
        mode:
          type: string
          enum: [read_only, read_write]
        indexEnabled:
          type: boolean
        allowPublicShares:
          type: boolean
    DeleteAdminMountResponse:
      type: object
      required: [id, deleted, deleteData, dataDeleted]
      properties:
        id:
          $ref: "#/components/schemas/OpaqueId"
        deleted:
          type: boolean
          const: true
        deleteData:
          type: boolean
        dataDeleted:
          type: boolean
...